--- server-app/src/app.js
@@ -1,11 +1,0 @@
-import express from 'express';
-import cors from 'cors';
-import authRoutes from './routes/authRoutes.js';
-
-const app = express();
-
-app.use(cors());
-app.use(express.json());
-app.use('/api', authRoutes);
-
-export { app };
