+++ server-app/src/index.js
@@ -1,16 +1,22 @@
-import { app } from './app.js';
-import { connectDB } from './config/db.js';
+import connectDB from './config/db.js';
+import express from 'express';
+import cors from 'cors';
+import authRoutes from './routes/authRoutes.js';
+import dotenv from 'dotenv';
 
-const PORT = process.env.PORT || 5000;
+dotenv.config();
 
-const startServer = async () => {
-    try {
-        await connectDB();
-        app.listen(PORT, () => console.log(`Server running on port ${PORT}`));
-    } catch (error) {
-        console.error('Error starting server:', error);
-        process.exit(1);
-    }
-};
+const app = express();
+const PORT = process.env.PORT || 3000;
 
-startServer();
+app.use(cors());
+app.use(express.json());
+app.use('/api/auth', authRoutes);
+
+app.listen(PORT, () => {
+    console.log('Server is running on port 3000');
+});
+
+connectDB();
+
+export default app;
