+++ server-app/src/routes/authRoutes.js
@@ -5,7 +5,7 @@ import { registerSchema, loginSchema } from '../schemas/auth.schema.js';
 
 const router = Router();
 
-router.post('/auth/register', validateSchema(registerSchema), register);
-router.post('/auth/login', validateSchema(loginSchema), login);
+router.post('/register', validateSchema(registerSchema), register);
+router.post('/login', validateSchema(loginSchema), login);
 
 export default router;
