--- server-app/src/server.js
@@ -1,18 +1,0 @@
-import { app } from './app.js';
-import { connectDB } from './config/db.js';
-
-const PORT = process.env.PORT || 3000;
-
-const startServer = async () => {
-  try {
-    await connectDB();
-    app.listen(PORT, () => {
-      console.log(`Servidor corriendo en puerto ${PORT}`);
-    });
-  } catch (error) {
-    console.error('Error al iniciar el servidor:', error);
-    process.exit(1);
-  }
-};
-
-startServer();
